feat: enhance client session management and configuration
- Added ClientSessionIdentifier and ClientSessionState for better session handling.
- Updated Client struct to use TlsStream for secure connections.
- Introduced ClientCertificateVerifier for client certificate validation.
- Refactored ClientRepository to manage clients using ClientSessionIdentifier.
- Enhanced ClientOptions with default values and mutators.
- Implemented message handling for various message types in the messages module.
- Improved configuration loading with serde and config crate.
- Added constants for maximum node and session IDs.
- Updated main function to initialize logging and server configuration.
- Added new message handlers for ACL, BanList, ChannelState, and others.

+++ src/types.rs
@@ -1,1 +1,1 @@
-pub type NodeIdentifier = i64;
+pub type NodeIdentifier = u16;
